* changed Rampart names based on amount of adds
