+++ scripts/zones/Arrapago_Remnants/mobs/Archaic_Rampart_2.lua
@@ -1,0 +1,59 @@
+-----------------------------------
+-- Area: Archaic Rampart
+--  MOB: Arrapago Remnants
+-----------------------------------
+mixins = {require("scripts/mixins/families/rampart")}
+local ID = require("scripts/zones/Arrapago_Remnants/IDs")
+require("scripts/globals/instance")
+require("scripts/globals/status")
+-----------------------------------
+function onMobSpawn(mob)
+    local instance = mob:getInstance()
+    mob:setLocalVar("swapTime", os.time())
+end
+
+function onMobFight(mob,target)
+    local instance = mob:getInstance()
+    local popTime = mob:getLocalVar("lastPetPop")
+    local POS = mob:getPos()
+    local PET1 = instance:getEntity(bit.band((mob:getID() +1), 0xFFF), dsp.objType.MOB)
+    local PET2 = instance:getEntity(bit.band((mob:getID() +2), 0xFFF), dsp.objType.MOB)
+
+    if os.time() - popTime > 15 then
+        if (not PET1:isSpawned()) then
+            PET1:setSpawn(POS.x, POS.y, POS.z, POS.rot)
+            mob:useMobAbility(2034)
+            mob:setLocalVar("lastPetPop", os.time())
+            mob:timer(2500, function(mob)
+            SpawnMob((mob:getID() +1), instance)
+            end)
+        elseif (not PET2:isSpawned()) then
+            PET2:setSpawn(POS.x, POS.y, POS.z, POS.rot)
+            mob:useMobAbility(2034)
+            mob:setLocalVar("lastPetPop", os.time())
+            mob:timer(2500, function(mob)
+            SpawnMob((mob:getID() +2), instance)
+            end)
+        end
+    end
+    if PET1:isSpawned() then
+        PET1:updateEnmity(target)
+    end
+    if PET2: isSpawned() then
+        PET1:updateEnmity(target)
+    end
+end
+
+function onMobDeath(mob, player, isKiller)
+    local instance = mob:getInstance()
+    if (ID.mob[6].rampart1 == mob:getID() or ID.mob[6].rampart2 == mob:getID()) then
+        if instance:getStage() == 6 and instance:getProgress() >= 1 then
+            if (isKiller) then
+                instance:setProgress(instance:getProgress() + 1)
+            end
+        end
+    end
+end
+
+function onMobDespawn(mob)
+end
